+++ docs/quickstart/assessment/standalone-assessment.php
@@ -1,4 +1,5 @@
 <?php
+
     /**
      * Copyright (c) 2021 Learnosity, MIT License
      *
@@ -60,9 +61,10 @@
         'items',
         $security,
         $consumerSecret,
-        $request);
+        $request
+    );
     $initOptions = $init->generate(); // JSON blob of signed config params.
-?>
+    ?>
 
 <!-- Section 2: Web page content. -->
 <!DOCTYPE html>